# process: replace strconv.ParseUint in parseMappings