@rockdaboot 2025-07-21
Would be interesting if you could compare benchmarks with https://github.com/tmthrgd/go-hex