@florianl 2025-07-21
> Would be interesting if you could compare benchmarks with https://github.com/tmthrgd/go-hex

I don't think this is a fair comparision, as the named package uses assembly code and only works for amd64.
As this is a fair simple change, I try to avoid adding unnecessary dependencies and also make sure the solution works on all supported architectures.