@rockdaboot 2025-07-21
> > Would be interesting if you could compare benchmarks with https://github.com/tmthrgd/go-hex
> 
> I don't think this is a fair comparision, as the named package uses assembly code and only works for amd64. As this is a fair simple change, I try to avoid adding unnecessary dependencies and also make sure the solution works on all supported architectures.

If I read the code correctly, it uses SIMD for amd64 and falls back to non-SIMD for other architectures. It should work everywhere.
You said, that "a significant part of process.parseMappings is spent in strconv.ParseUint". So it makes sense to at least know how much we could shave off. Maybe it's worth to add another dependency, maybe not. But we should at least have some numbers.