@christos68k 2025-07-21
I'm ambivalent about this, according to the flamegraph most of the time is spent in `Scan` which makes me wonder if the the ~20% improvement in the time `ParseUint` takes is worth the special-casing we do here and the corner cases that come with it.