@florianl 2025-07-22
The flamegraph was recorded with `perf record --freq=max ...`. So there is a higher precision than the 20Hz of profiling data, OTel eBPF profiler generates.

> What is the system workload for this?

Google Chrome, devfiler, OTel eBPF profiler, perf and Elastics security applications - nothing special. I would describe it as regular workload with a total system load of 20% to 25%.

> During startup we read lot of mappings and this likely shows up. But is this important after profiler startup? 

The recorded process, OTel eBPF profiler, was running for (+2) hours at this point. The flamegraph does not show the startup and initializen phase.