@christos68k 2025-07-22
I also looked at one of our K8s environments where `pf-host-agent` as whole is using 1-3% CPU with `ElasticSearch` dominating. The amount of CPU time spent in strconv.ParseUint is less than 0.02% of the 1-3% so I agree with @rockdaboot that this optimization is not worth the extra code/corner cases.